+++ packages/core/src/model-types.ts
@@ -56,8 +56,5 @@ export type AwsBedrockChatModel = LiteralUnion<
  * AI Core open source models for chat completion.
  */
 export type AiCoreOpenSourceChatModel = LiteralUnion<
-  | 'mistralai--mistral-large-instruct'
-  | 'mistralai--mistral-small-instruct'
-  | 'alephalpha-pharia-1-7b-control'
-  | 'deepseek-ai--deepseek-r1'
+  'mistralai--mistral-large-instruct' | 'mistralai--mistral-small-instruct'
 >;