+++ CHANGELOG.md
@@ -28,6 +28,15 @@
 
 -
 
+# 2.1.0
+## Fixed Issues
+
+- [orchestration] Fix JSDoc example of `buildAzureContentSafetyFilter()` function. (0cf7d80)
+
+## Improvements
+
+- [core] Remove `alephalpha-pharia-1-7b-control` and `deepseek-ai--deepseek-r1` from available model list. (0cf7d80)
+
 # 2.0.0
 ## Compatibility Notes