+++ sample-cap/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ai-sdk/sample-cap
 
+## 2.1.0
+
+### Patch Changes
+
+- Updated dependencies [0cf7d80]
+  - @sap-ai-sdk/orchestration@2.1.0
+  - @sap-ai-sdk/ai-api@2.1.0
+  - @sap-ai-sdk/foundation-models@2.1.0
+
 ## 2.0.0
 
 ### Patch Changes